made modules in JS
