+++ index.html
@@ -27,6 +27,6 @@
     </header>
     <main></main>
     <footer></footer>
-    <script src="/src/JS/header.js"></script>
+    <script type="module" src="/src/JS/main.js"></script>
   </body>
 </html>
