+++ src/JS/main.js
@@ -1,0 +1,2 @@
+import { mobileMenu } from "./mobilmenu.js";
+mobileMenu();
